Finish editing form

+++ Register/models.py
@@ -9,7 +9,7 @@ class Student(models.Model):
     father_name = models.CharField(max_length=50, verbose_name="نام پدر")
     national_code = models.CharField(max_length=10, unique=True, verbose_name="کد ملی")
     birth_date = models.DateField(verbose_name="تاریخ تولد")
-    payment_amount = models.FloatField(blank=True, null=True, default=0, verbose_name="مبلغ پرداخت شده")
+    payment_amount = models.IntegerField(blank=True, null=True, default=0, verbose_name="مبلغ پرداخت شده")
     contact_number = models.CharField(max_length=15, verbose_name="تلفن تماس")
     classes = models.ManyToManyField('Class', related_name='students', verbose_name="کلاس‌ها")
 
@@ -19,7 +19,7 @@ class Student(models.Model):
 
 class Class(models.Model):
     class_name = models.CharField(max_length=100, verbose_name="نام کلاس")
-    class_cost = models.FloatField(blank=True, null=True, default=0, verbose_name="هزینه کلاس")
+    class_cost = models.IntegerField(blank=True, null=True, default=0, verbose_name="هزینه کلاس")
 
     def __str__(self):
         return self.class_name
